Shader + NEED FIX

+++ Shader.cpp
@@ -19,6 +19,7 @@ Shader::Shader()
 		glGetShaderInfoLog(vertexShader, 512, NULL, infoLog);
 		LOGC("ERROR::SHADER::VERTEX::COMPILATION_FAILED\n %s", infoLog);
 	}
+	
 	// fragment shader
 	int fragmentShader = glCreateShader(GL_FRAGMENT_SHADER);
 	glShaderSource(fragmentShader, 1, &fragmentShaderSource, NULL);
@@ -30,6 +31,7 @@ Shader::Shader()
 		glGetShaderInfoLog(fragmentShader, 512, NULL, infoLog);
 		LOGC("ERROR::SHADER::FRAGMENT::COMPILATION_FAILED\n %s", infoLog);
 	}
+
 	// link shaders
 	ID = glCreateProgram();
 	glAttachShader(ID, vertexShader);
@@ -41,6 +43,7 @@ Shader::Shader()
 		glGetProgramInfoLog(ID, 512, NULL, infoLog);
 		LOGC("ERROR::SHADER::PROGRAM::LINKING_FAILED\n %s", infoLog);
 	}
+	
 	glDeleteShader(vertexShader);
 	glDeleteShader(fragmentShader);
 }
@@ -51,6 +54,7 @@ void Shader::use()
 	glEnable(GL_BLEND);
 }
 
+
 void Shader::stop()
 {
 	glUseProgram(0);
@@ -59,7 +63,7 @@ void Shader::stop()
 
 void Shader::setBool(const std::string &name, bool value) const
 {
-	glUniform1i(glGetUniformLocation(ID, name.c_str()), (int)value);
+	glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
 }
 
 void Shader::setInt(const std::string &name, int value) const
